requirements.txt
api==0.0.7
APScheduler==3.7.0
asgiref==3.4.0
asyncio==3.4.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.5
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.1
django-heroku==0.3.1
django-nine==0.2.4
django-redis==5.0.0
djangorestframework==3.12.4
elasticsearch==7.13.1
elasticsearch-dsl==7.3.0
gevent==21.1.2
google-api-core==1.30.0
google-api-python-client==2.11.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
greenlet==1.1.0
grequests==0.6.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
nose==1.3.7
oauthlib==3.1.1
packaging==21.0
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.2.0
zope.event==4.5.0
zope.interface==5.4.0
